requirements_production.txt
# Flask e extensões
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Mail==0.10.0
Flask-SocketIO==5.3.6
Flask-Compress==1.14
Flask-Limiter==3.5.0

# Banco de dados
psycopg2-binary==2.9.7
redis==4.6.0

# Autenticação e segurança
passlib[argon2]==1.7.4
cryptography==41.0.4

# WebSocket e async
eventlet==0.33.3
gevent==23.7.0
gevent-websocket==0.10.1

# Processamento de imagens
Pillow==10.0.1

# Utilitários
python-slugify==8.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7

# Monitoramento e logs
prometheus-client==0.17.1
structlog==23.1.0

# Backup e manutenção
schedule==1.2.0

# Produção
gunicorn==21.2.0
supervisor==4.2.5


